Move LengthLimitedWriter into the jackson module
YamlCodec.toYamlWithLengthLimit needs the same length-limited writer that
JsonCodec.toJsonWithLengthLimit uses, so the class and its tests move out
of json into the shared jackson module and the class becomes public.
